# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/product/": {
            "get": {
                "description": "Get all products",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all Products",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/add": {
            "post": {
                "description": "Add a single product",
                "produces": [
                    "application/json"
                ],
                "summary": "Add Product",
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/product/addMany": {
            "post": {
                "description": "Add Many product at once",
                "produces": [
                    "application/json"
                ],
                "summary": "Add Many Product",
                "responses": {
                    "201": {
                        "description": "created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reports/sales-by-brand": {
            "get": {
                "description": "Get Sales by product brand",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Reports Sales by brand",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reports/sales-by-product": {
            "get": {
                "description": "Get Sales by product reports",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Reports Sales by product",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reports/summery": {
            "get": {
                "description": "Get Summery Reports",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Summery Reports",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/reports.SummeryReport"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sales": {
            "get": {
                "description": "Get All Sales",
                "produces": [
                    "application/json"
                ],
                "summary": "Get All Sales",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "reports.SummeryReport": {
            "type": "object",
            "properties": {
                "dateOfHighestSales": {
                    "type": "string"
                },
                "dateOfLeastSales": {
                    "type": "string"
                },
                "leastProfitableProduct": {
                    "type": "string"
                },
                "mostProfitableProduct": {
                    "type": "string"
                }
            }
        }
    }
}